Base for project written
I've made sure that Main can now be compiled, although it will crash if you attempt to do anything useful with it. This is a large commit, but given that no-one appears to have been working on this so far apart from me, it shouldn't matter. We'll assign individual tasks to get this project off the ground soonish.