We now have output with Cassava!
I'm not going to do output for the simple one, it drives me up the wall to do the parsing. Versus a function that does everything for me? The choice is obvious. I had to remove all my packages though and reset to get this to work, but oh well.