Adding ToNamedRecord instance for OrderBookEntry
Forgot to add this to the writer, which means that there would be no actual output (as there is no instance). I haven't yet put in the code needed to turn the monstrosity of a transaction element into usable blocks yet, but that will happen gradually.

Once that is done, then all that remains is adding entryTime, oldPrice and oldVolume.