Stopping money overflows in Trader
Since we assume that we can always sell everything when we want to,
this leads to some unexpected effects when sell above market value.
Like having more money than God himself. So I've changed the fancy
based off the difference thing and we now just have a constant that
gets multiplied. This makes the code rather fast though, so not too
bad a thing.