Adding a TODO to help finalize the design
I'm going to take a break from doing development on this while I
work on all my other assignments. This should bring you up to speed
with what I want to do next with the current code base.

We should be focusing on getting the code in a workable state and then
work on optimizing the functions later.