tran is Maybe TransId, cleanup, process trades
Now we can correctly parse fields that don't have A or B in the line.
This means that we are now effectively back where we started in terms
of quality of data parsing and output. I have linked the trade matcher
that Sam wrote into the update method and I removed old code from Types
that currently isn't being used.