Added createStrategy and placeholders for buy/sell
createStrategy is going to be the wrapper function for our
strategy generator to help increase legibility. I've chosen
the names "bestBuy" and "bestSell" to denote our strategy
to buy at the best time and sell at the best price currently known.

"bestBuy" and "bestSell" currently default to nothing, which is going
to be changed when both functions are implemented.