Removed old buy, sell is complete, fixed new buy
*Had problems making partial patches.

I've cleaned up the actors section by removing the old buyer.
This required some cleanup in the top of the module as well.

To sell, we just put the share that we want to sell back into
circulation. OrderBook is going to magically know that it's our
trade and give us the money accordingly.

Buy could technicially buy more than was available. Now it can only
buy as much as the share holds.