Made stubs for trader interaction, trade matches
With this I should hopefully be on the right path to system completion.
We currently do not handwave away the actual trade matching, but we
need some carefully curated values to perform actions upon.

That's going to require changing the data structure and the update
methods that get called on new data (enter, amend, delete).